Conflicts:
	src/main/webapp/jsp/interface/header_relate.jsp

+++ src/main/java/com/dc/esb/servicegov/controller/ServiceController.java
@@ -36,6 +36,8 @@ public class ServiceController {
     @Autowired
     private ServiceServiceImpl serviceServiceImpl;
     @Autowired
+    private OperationServiceImpl operationServiceImpl;
+    @Autowired
     private ServiceCategoryServiceImpl serviceCategoryServiceImpl;
     @Autowired
     private ProcessContextServiceImpl processContextService;
@@ -129,6 +131,10 @@ public class ServiceController {
     public
     @ResponseBody
     boolean deleteService(@PathVariable String Id) {
+        Map<String,String> map = new HashMap<String, String>();
+        map.put("serviceId",Id);
+        List<Operation> list = operationServiceImpl.findBy(map);
+        if(list.size() > 0) return false;
         OperationLog operationLog = systemLogService.record("服务","删除","服务ID：" +Id);
 
         serviceServiceImpl.deleteById(Id);
@@ -168,6 +174,9 @@ public class ServiceController {
     public
     @ResponseBody
     boolean deleteServiceCategory(@PathVariable String Id) {
+        //大类下是否有服务
+        List<Service> list = serviceServiceImpl.findBy("categoryId", Id);
+        if(list.size()>0) return false;
         OperationLog operationLog = systemLogService.record("服务分类", "删除", "分类ID：" + Id);
 
         serviceCategoryServiceImpl.deleteById(Id);
